first full data out and load working
