+++ oneQbit_2actions/hyperparams.py
@@ -1,0 +1,6 @@
+import os
+
+folder_name = "/oneQbit_2actions/"
+cwd = os.getcwd()
+results_dir = cwd + folder_name + "training_results/"
+
